Add Assets Folder (except Imports)
This is the Assets folder as of August 4, 2020 4:12 pm. It is all of the files I am using currently, minus the Imported folder which included assets  I previously used and the OVR  folder. In order for this to work you must install the Oculus Integration (https://assetstore.unity.com/packages/tools/integration/oculus-integration-82022) which is free on the Unity asset Store.